JENKINS-10073: Set ACCUREV_STREAM by parameter
An external entity provides the stream which the AccuRev plugin would use.

For example there is a parametrized job which's string parameter gets the name that the plugin should use.

Otherwise, an overwrite of %ACCUREV_STREAM% (e.g. set a parameter called ACCUREV_STREAM) would be an alternative.

It should be done like the "Build selector for Copy Artifact" which can be set as a parameter and the plugin is able to retrieve the value from this special parameter.

See: https://wiki.jenkins-ci.org/display/JENKINS/Copy+Artifact+Plugin

-----------

files contain fix to this improvement
usage:

*	make parameter in job e.g. acstream
*	then set in stream name field: ${acstream}
*	SCM polling will go against the default value of the parameter

One thing what we didn’t put in there is the catch of Choice parameters for the SCM polling since we were happy with the simple String parameter.

Signed-off-by: helterscelter <https://github.com/helterscelter>